axi_dmac: diagnostic interface
This change adds a diagnostic interface to the DMAC core.
The interface exposes internal information about the core,
information which can't be exposed through AXI registers
due the latency and update rate.

Such information is the fullness of the internal buffer.
For flexibility this is exposed in bursts and beats aswell
with reduced resource utilization due the reuse of the
existing beat counters.
The exposed signals are driven from the destination clock domain,
as this is reflected in their name.

The signals have a fixed size and they are dimensioned by
taking in account the supported maximum number of bursts of 128 and
burst length of 256 beats.